refactor: extract connector utils and schemas from connectors.ts
Split the 4199-line connectors.ts into three focused modules:
- connectors.ts (3764 lines): type definitions + connector data
- connector-utils.ts (344 lines): 13 utility functions
- connector-schemas.ts (103 lines): 7 Zod response schemas

Updates index.ts re-exports and fixes 4 direct import paths.
No behavioral changes — pure code movement.

Closes #7833
