+++ turbo/packages/core/src/contracts/__tests__/firewall-secret-consistency.test.ts
@@ -3,7 +3,7 @@ import { connectorTypeSchema } from "../connectors";
 import {
   getConnectorEnvironmentMapping,
   getConnectorAuthMethods,
-} from "../connectors";
+} from "../connector-utils";
 import { getConnectorFirewall, isFirewallConnectorType } from "../../firewalls";
 
 /**
